Renaming
Renamed PinnableMixin to PinnableModelMixin, PinnablesMixin to PinnablesViewMixin, and pins.py to pin_models.py
